renamed music-album.rb to snake_case music_album.rb
