+++ genre.rb
@@ -1,6 +1,6 @@
 class Genre
   attr_accessor :id, :name, :items
-  
+
   def initialize(name)
     @id = id || Random.rand(1..9999)
     @name = name
